[enh #96] 作業必要ファイル追加
