+++ ProjectAzur/ProjectAzur/MainGame/GameScene/Player/PlayerShape/paPlayerFishShell.cpp
@@ -1,0 +1,27 @@
+﻿//==================================================================================================================================//
+//!< @file		paPlayerFishShell.cpp
+//!< @brief		pa::PlayerFishShellクラスの実装
+//!< @author	T.Haga
+//!< @data		作成日時：2018/01/21	更新履歴：
+//==================================================================================================================================//
+
+/* Includes --------------------------------------------------------------------------------------------------- */
+
+#include "paPlayerFishShell.h"
+
+namespace pa
+{
+
+/* Public Functions ------------------------------------------------------------------------------------------- */
+
+PlayerFishShell::PlayerFishShell()
+{}
+
+PlayerFishShell::~PlayerFishShell()
+{}
+
+}	// namespace pa
+
+//==================================================================================================================================//
+// END OF FILE
+//==================================================================================================================================//
