+++ ProjectAzur/ProjectAzur/MainGame/GameScene/Player/PlayerShape/paPlayerFishShell.h
@@ -1,0 +1,34 @@
+﻿//==================================================================================================================================//
+//!< @file		paPlayerFishShell.h
+//!< @brief		pa::PlayerFishShellクラスのヘッダ
+//!< @author	T.Haga
+//!< @data		作成日時：2018/01/21	更新履歴：
+//==================================================================================================================================//
+
+#ifndef PA_PLAYER_FISH_SHELL_H
+#define PA_PLAYER_FISH_SHELL_H
+
+namespace pa
+{
+
+//===================================================================================//
+//!< プレイヤーの魚の殻のクラス 
+//===================================================================================//
+class PlayerFishShell
+{
+public:
+	/** Constructor */
+	PlayerFishShell();
+
+	/** Destructor */
+	~PlayerFishShell();
+
+};	// class PlayerFishShell
+
+}	// namespace pa
+
+#endif	// PA_PLAYER_FISH_SHELL_H
+
+//==================================================================================================================================//
+// END OF FILE
+//==================================================================================================================================//
