+++ ProjectAzur/ProjectAzur/MainGame/GameScene/Player/PlayerShape/paPlayerOriginalShape.cpp
@@ -1,0 +1,27 @@
+﻿//==================================================================================================================================//
+//!< @file		paPlayerOriginalShape.cpp
+//!< @brief		pa::PlayerOriginalShapeクラスの実装
+//!< @author	T.Haga
+//!< @data		作成日時：2018/01/21	更新履歴：
+//==================================================================================================================================//
+
+/* Includes --------------------------------------------------------------------------------------------------- */
+
+#include "paPlayerOriginalShape.h"
+
+namespace pa
+{
+
+/* Public Functions ------------------------------------------------------------------------------------------- */
+
+PlayerOriginalShape::PlayerOriginalShape()
+{}
+
+PlayerOriginalShape::~PlayerOriginalShape()
+{}
+
+}	// namespace pa
+
+//==================================================================================================================================//
+// END OF FILE
+//==================================================================================================================================//
