+++ ProjectAzur/ProjectAzur/MainGame/GameScene/Player/PlayerShape/paPlayerOriginalShape.h
@@ -1,0 +1,35 @@
+﻿//==================================================================================================================================//
+//!< @file		paPlayerOriginalShape.h
+//!< @brief		pa::PlayerOriginalShapeクラスのヘッダ
+//!< @author	T.Haga
+//!< @data		作成日時：2018/01/21	更新履歴：
+//==================================================================================================================================//
+
+#ifndef PA_PLAYER_ORIGINAL_SHAPE_H
+#define PA_PLAYER_ORIGINAL_SHAPE_H
+
+namespace pa
+{
+
+//===================================================================================//
+//!< プレイヤーのオリジナルの形状クラス
+//===================================================================================//
+class PlayerOriginalShape
+{
+
+public:
+	/** Constructor */
+	PlayerOriginalShape();
+
+	/** Destrutor */
+	~PlayerOriginalShape();
+
+};	// class PlayerOriginalShape
+
+}	// namespace pa
+
+#endif	// PA_PLAYER_ORIGINAL_SHAPE_H
+
+//==================================================================================================================================//
+// END OF FILE
+//==================================================================================================================================//
