+++ ProjectAzur/ProjectAzur/ProjectAzur.vcxproj
@@ -135,7 +135,9 @@
     <ClCompile Include="MainGame\GameScene\Player\paPlayer.cpp" />
     <ClCompile Include="MainGame\GameScene\Player\paPlayerAction.cpp" />
     <ClCompile Include="MainGame\GameScene\Player\paPlayerCollider.cpp" />
-    <ClCompile Include="MainGame\GameScene\Player\paPlayerShape.cpp" />
+    <ClCompile Include="MainGame\GameScene\Player\PlayerShape\paPlayerFishShell.cpp" />
+    <ClCompile Include="MainGame\GameScene\Player\PlayerShape\paPlayerOriginalShape.cpp" />
+    <ClCompile Include="MainGame\GameScene\Player\PlayerShape\paPlayerShape.cpp" />
     <ClCompile Include="MainGame\GameScene\Stage\Fish\FishCollider\paFishCollider.cpp" />
     <ClCompile Include="MainGame\GameScene\Stage\Fish\FishCreator\paFishShapeCreator.cpp" />
     <ClCompile Include="MainGame\GameScene\Stage\Fish\FishCreator\psFishCreator.cpp" />
@@ -177,7 +179,9 @@
     <ClInclude Include="MainGame\GameScene\Player\paPlayerAction.h" />
     <ClInclude Include="MainGame\GameScene\Player\paPlayerCollider.h" />
     <ClInclude Include="MainGame\GameScene\Player\paPlayerCommonDeclaration.h" />
-    <ClInclude Include="MainGame\GameScene\Player\paPlayerShape.h" />
+    <ClInclude Include="MainGame\GameScene\Player\PlayerShape\paPlayerFishShell.h" />
+    <ClInclude Include="MainGame\GameScene\Player\PlayerShape\paPlayerOriginalShape.h" />
+    <ClInclude Include="MainGame\GameScene\Player\PlayerShape\paPlayerShape.h" />
     <ClInclude Include="MainGame\GameScene\Stage\Fish\FishCollider\paFishCollider.h" />
     <ClInclude Include="MainGame\GameScene\Stage\Fish\FishCreator\paFishCreator.h" />
     <ClInclude Include="MainGame\GameScene\Stage\Fish\FishCreator\paFishShapeCreator.h" />
